Move parse.rs to mount.rs
In anticipation of adding parallel support for mountinfo, move support
for parsing /proc/mounts (and /etc/fstab) to mount.rs.

Modify examples as needed.
