+++ src/lib.rs
@@ -13,7 +13,6 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 pub use error::*;
-pub use parse::*;
 
 mod error;
-mod parse;
+pub mod mount;
